ADD: admin page & admin navbar
