+++ projects/client/src/components/Navbar/index.js
@@ -1,11 +1,12 @@
 import { useState } from "react"
 import { AnimatePresence, motion } from "framer-motion"
 import Modal from "../Modal"
-import NavMenu from "./menu"
+import UserNavMenu from "./user.menu"
+import AdminNavMenu from "./admin.menu";
 
-export default function Navbar ({ user, isLogin, setIsLogin }) {
+export default function Navbar ({ user, isLogin, setIsLogin, isSidebarActive, setIsSidebarActive }) {
     const [showModal, setShowModal] = useState({ show: false, context: "" })
-  
+
     const handleShowModal = (context) => {
         setShowModal({ show: true, context })
         document.body.style.overflow = "hidden"
@@ -26,7 +27,7 @@ export default function Navbar ({ user, isLogin, setIsLogin }) {
                     className="nav"
                 >
                     <div className="navbar container flex">
-                        <div className="lg:block ml-10">
+                        <div className="lg:block">
                             <a 
                                 href="/" 
                                 className="flex items-center gap-2 text-2xl font-bold tracking-tighter text-dark"
@@ -35,15 +36,23 @@ export default function Navbar ({ user, isLogin, setIsLogin }) {
                             </a>
                         </div>
 
-                        <div className={`nav-menu-wrapper justify-end  ${isLogin ? "lg:w-1/3" :""}`} >
-
-                        <NavMenu
+                    {user.role === 1?
+                        <AdminNavMenu
+                            isLogin={isLogin}
+                            setIsLogin={setIsLogin}
+                            user={user}
+                            isSidebarActive={isSidebarActive}
+                            setIsSidebarActive={setIsSidebarActive}
+                        />
+                        :
+                        <UserNavMenu
                             isLogin={isLogin}
                             setIsLogin={setIsLogin}
                             handleShowModal={handleShowModal}
                             user={user}
                         />
-                        </div>
+                        }   
+                    
                     </div>
 
                     <Modal
